#11 use only PUT method to manipulate grades

+++ src/main/java/com/mitrais/bootcamp/rms/data/controller/GradeController.java
@@ -5,9 +5,9 @@ import com.mitrais.bootcamp.rms.data.entity.Grade;
 import com.mitrais.bootcamp.rms.data.repository.EmployeeRepository;
 import com.mitrais.bootcamp.rms.data.repository.GradeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashSet;
 import java.util.NoSuchElementException;
 import java.util.Set;
 
@@ -26,42 +26,47 @@ public class GradeController {
     protected GradeController() {
     }
 
-    @RequestMapping(method = RequestMethod.POST)
-    @ResponseStatus(HttpStatus.CREATED)
-    public void addGrade(@PathVariable(value = "empId") long empId, @RequestBody Grade grade) {
-        Employee emp = verifyEmployee(empId);
-
-        emp.addGrade(grade);
-
-        employeeRepository.save(emp);
-    }
-
     @RequestMapping(method = RequestMethod.GET)
     public Set<Grade> getGrades(@PathVariable(value = "empId") long empId) {
         Employee emp = verifyEmployee(empId);
         return emp.getGrades();
     }
 
-    @RequestMapping(method = RequestMethod.DELETE, path = "/{gradeId}")
-    public void delete(@PathVariable(value = "empId") long empId, @PathVariable(value = "gradeId") String gradeId) {
-        Employee emp = verifyEmployee(empId);
-        Grade grade = verifyGrade(gradeId);
-
-        emp.removeGrade(grade);
-        gradeRepository.delete(grade);
-        employeeRepository.save(emp);
-    }
-
     @RequestMapping(method = RequestMethod.PUT)
-    public Grade update(@PathVariable(value = "empId") int empId, @RequestBody Grade newGrade) {
+    public Set<Grade> update(@PathVariable(value = "empId") int empId, @RequestBody Set<Grade> newGrades) {
         Employee emp = verifyEmployee(empId);
-        Grade grade = verifyGrade(newGrade.getGradeId());
-        grade.setGrade(newGrade.getGrade());
-        grade.setStartDate(newGrade.getStartDate());
-        grade.setEndDate(newGrade.getEndDate());
-        grade.setDs(newGrade.getDs());
 
-        return gradeRepository.save(grade);
+        Set<Grade> oldGrades = new HashSet<Grade>(emp.getGrades());
+
+        for (Grade grade: oldGrades) {
+            // handle delete
+            if (!newGrades.contains(grade)) {
+                emp.removeGrade(grade);
+                gradeRepository.delete(grade);
+            } else {
+                // handle update
+                for (Grade newGrade : newGrades) {
+                    if (newGrade.equals(grade)) {
+                        grade.setGrade(newGrade.getGrade());
+                        grade.setStartDate(newGrade.getStartDate());
+                        grade.setEndDate(newGrade.getEndDate());
+                        grade.setDs(newGrade.getDs());
+
+                        gradeRepository.save(grade);
+                    }
+                }
+            }
+        }
+
+        // handle add
+        for (Grade newGrade: newGrades) {
+            if (!oldGrades.contains(newGrade)) {
+                emp.addGrade(newGrade);
+            }
+        }
+
+        employeeRepository.save(emp);
+        return newGrades;
     }
 
     private Employee verifyEmployee(long empId) throws NoSuchElementException {
@@ -70,13 +75,5 @@ public class GradeController {
             throw new NoSuchElementException("Employee does not exist " + empId);
         }
         return employee;
-    }
-
-    private Grade verifyGrade(String gradeId) throws NoSuchElementException {
-        Grade grade = gradeRepository.findByGradeId(gradeId);
-        if (grade == null) {
-            throw new NoSuchElementException("Grade does not exist " + gradeId);
-        }
-        return grade;
     }
 }
